# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SCORM Engine API
  version: 0.1.0
  summary: Modern SCORM runtime + tracking engine for TypeScript applications.
  description: |
    The SCORM Engine API is a **vendor-neutral REST contract** for SCORM 1.2 and
    SCORM 2004 course ingestion, runtime tracking, and analytics. It is the
    primary contract for our client SDK (`scorm-engine-client`), but the API is
    designed to be consumed by **any HTTP client**.

    The engine ships static-asset routes (`/play/:attemptId`, `/runtime.js`) that
    bridge browser-side SCOs to the API. The runtime bridge is a **clean-room,
    spec-compliant implementation of the SCORM JavaScript API** — it has no
    third-party SCORM dependencies. Because it follows the spec, any SCORM-compliant
    course works regardless of the authoring tool or API wrapper it was built with.

    ## Versioning
    The API is versioned via the URL prefix (`/api/v1`). Breaking changes ship under
    a new prefix. Additive changes (new optional fields, new endpoints) are non-breaking.

    ## Authentication
    Two auth modes:
    - **API key** (`X-API-Key` header): service-to-service calls from an LMS backend.
    - **Attempt JWT** (`Authorization: Bearer <jwt>`): short-lived token scoped to
      a single attempt, used by the in-browser player. Issued by `/auth/attempt-token`.

    ## Tenancy
    Every resource is scoped to a tenant resolved from the API key. Attempt JWTs
...
  contact:
    name: SCORM Engine
...
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /play/{attemptId}:
    parameters:
      - $ref: '#/components/parameters/AttemptId'
    get:
      tags: [Embed]
      summary: Self-contained iframe player page
      description: |
        Returns an HTML page that loads the SCO, injects the SCORM runtime API
        (`window.API` / `window.API_1484_11`), and persists progress to this engine.
        Host with a single `<iframe>` — no SDK or framework required.
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
          description: Attempt JWT (issued by `/auth/attempt-token`).
        - name: theme
          in: query
          schema: { type: string, enum: [light, dark, auto], default: auto }
        - name: source
          in: query
          description: |
            Where the SCO content comes from. `engine` (default) serves the
            uploaded package via `/courses/{id}/content/*` (and sets an auth
            cookie so sub-resources authenticate). `external` embeds the URL in
            `src` instead.
          schema: { type: string, enum: [engine, external], default: engine }
        - name: src
          in: query
          description: Absolute SCO entry URL. Required when `source=external`.
          schema: { type: string, format: uri }
        - name: scoId
          in: query
          schema: { type: string }
          description: Optional — override the default SCO (engine source only).
      responses:
        '200':
          description: HTML player page
          content:
            text/html:
              schema: { type: string }
        '400':
          description: Missing content source (e.g. source=external without src).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Token is not valid for this attempt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /runtime.js:
    get:
      tags: [Embed]
      summary: Standalone SCORM runtime bridge script
      description: |
        Returns a small JS file that, when included via `<script>`, exposes
        `window.API` (SCORM 1.2) and `window.API_1484_11` (SCORM 2004) wired to
        this engine. A clean-room, spec-compliant implementation — no third-party
        SCORM libraries are bundled or required. The SCO inside an `<iframe>`
        finds the globals via the normal SCORM lookup walk (parent, then opener)
        and behaves exactly as it would against any conformant LMS.

        ```html
        <script src="https://your-engine/runtime.js?attemptId=...&token=..."></script>
        <iframe src="./course/index.html"></iframe>
        ```
      security: []
      parameters:
        - name: attemptId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: token
          in: query
          required: true
          schema: { type: string }
          description: Attempt JWT (issued by `/auth/attempt-token`).
      responses:
        '200':
          description: JavaScript runtime bridge
          content:
            application/javascript:
              schema: { type: string }
            text/javascript:
              schema: { type: string }

  # ─────────────────────────────────────────────────────────────────────────
  # Auth
  # ─────────────────────────────────────────────────────────────────────────
  /auth/attempt-token:
    post:
      tags: [Auth]
      summary: Issue a short-lived JWT scoped to a single attempt
      description: |
        Called by the LMS backend to mint a token the browser-side player can use
        without exposing the long-lived API key. Token TTL defaults to the attempt
        duration limit (configurable, max 24h).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attemptId]
              properties:
                attemptId: { type: string, format: uuid }
                ttlSeconds: { type: integer, minimum: 60, maximum: 86400, default: 3600 }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expiresAt: { type: string, format: date-time }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────────────────────────────────────────────────────────
  # Courses
  # ─────────────────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Long-lived tenant API key. Used for server-to-server calls.
    AttemptJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT scoped to a single attempt. Issued via /auth/attempt-token.
        Used by the in-browser player.

  # ───────────────────────────────────────────────────────────────────────
  # Reusable parameters
  # ───────────────────────────────────────────────────────────────────────
  parameters:
...
    AttemptId:
      name: attemptId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  # ───────────────────────────────────────────────────────────────────────
  # Schemas
  # ───────────────────────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code (e.g. `course_not_found`).
        message: { type: string }
        details:
          type: object
          additionalProperties: true
        requestId: { type: string }
...